mcp-adk/pyproject.toml
[project]
name = "mcp-adk"
version = "0.1.0"
description = "Skeleton project to connect MCP Server using ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "google-adk>=1.15.1",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "litellm>=1.77.5",
    "llama-index-embeddings-ollama>=0.8.3",
    "mcp[cli]>=1.15.0",
]
